Add an optional search path for `../pytorch`
I'm trying to find a way to support, in Pyre1, dual development
modes where pytorch might be installed via site-packages but
alternatively might be installed from source in the same project.

If we try to use a regular `search_path` it won't work; on my
machine I don't have `../pytorch`, and Pyre will crash complaining
about a missing search path elment.

Luckily, though, we added an optional_search_path with exactly the
behavior we need a while back: Pyre1 will ignore it if the path doesn't
exist, and include it if it does.

Note that