Improve autotuning log messages.
Elaborate more on what's happening during autotuning so user can understands better.
Set the expectation in the beginning that autotuning will take a while.
For the progress bar, instead of using the rich backend of tqdm, we can use
rich itself, which allows removing the warning message,
"TqdmExperimentalWarning: rich is experimental/alpha", as well as
customize the progress bar better (colors, remove some of the clutter
with expected time, which is not accurate, and rounds/s speed).
If cache hits so we don't need to run autotuning, let the user know and
how the cache can be removed.
Overall, autotuner log messages are cleaner and elaborate more on what's
happening.