Update pyrefly pre-commit hook.
This is a follow-up to a reviewer request in
https://github.com/pytorch/helion/pull/1143 to use an isolated
environment for the pyrefly pre-commit hook. Unfortunately, I
don't think that's practical to do (see
https://github.com/facebook/pyrefly-pre-commit/issues/7#issuecomment-3595346638).
But in the process of looking into this, I ended up giving the
hook a better name and versioning system, which this PR updates
helion's pre-commit config to use.